CI-629 altered ukprn to long

+++ src/SFA.DAS.Apprenticeships.Api.Types/AssessmentOrgs/OrganisationSummary.cs
@@ -12,6 +12,6 @@ namespace SFA.DAS.Apprenticeships.Api.Types.AssessmentOrgs
 
         public List<Link> Links { get; set; }
 
-        public int? Ukprn { get; set; }
+        public long? Ukprn { get; set; }
     }
 }
